requirements.txt
Django==5.1.6
psycopg2 == 2.9.10
psycopg2-binary==2.9.10
djangorestframework==3.15.2
Pillow == 11.1.0
jwt == 1.3.1
djangorestframework-simplejwt == 5.5.0
